requirements.txt
python-chess>=1.9.0
requests>=2.28.0
matplotlib>=3.5.0
seaborn>=0.12.0
python-dotenv>=0.20.0